requirements.txt
streamlit
numpy
opencv-python
tensorflow-cpu